JAMES-2082 Extract MessageAttachmentRep and MessageWithoutAttachment to separated model
