+++ mailbox/cassandra/src/main/java/org/apache/james/mailbox/cassandra/mail/CassandraMessageDAO.java
@@ -42,7 +42,6 @@ import static org.apache.james.mailbox.cassandra.table.CassandraMessageV1Table.T
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.ByteBuffer;
-import java.util.Date;
 import java.util.List;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
@@ -51,14 +50,12 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import javax.inject.Inject;
-import javax.mail.Flags;
 import javax.mail.util.SharedByteArrayInputStream;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.apache.james.backends.cassandra.CassandraConfiguration;
 import org.apache.james.backends.cassandra.init.CassandraTypesProvider;
 import org.apache.james.backends.cassandra.utils.CassandraAsyncExecutor;
-import org.apache.james.mailbox.MessageUid;
 import org.apache.james.mailbox.cassandra.CassandraMessageId;
 import org.apache.james.mailbox.cassandra.table.CassandraMessageV1Table.Attachments;
 import org.apache.james.mailbox.cassandra.table.CassandraMessageV1Table.Properties;
@@ -67,13 +64,9 @@ import org.apache.james.mailbox.model.AttachmentId;
 import org.apache.james.mailbox.model.Cid;
 import org.apache.james.mailbox.model.ComposedMessageId;
 import org.apache.james.mailbox.model.ComposedMessageIdWithMetaData;
-import org.apache.james.mailbox.model.MailboxId;
-import org.apache.james.mailbox.model.MessageAttachment;
-import org.apache.james.mailbox.model.MessageId;
 import org.apache.james.mailbox.store.mail.MessageMapper.FetchType;
 import org.apache.james.mailbox.store.mail.model.MailboxMessage;
 import org.apache.james.mailbox.store.mail.model.impl.PropertyBuilder;
-import org.apache.james.mailbox.store.mail.model.impl.SimpleMailboxMessage;
 import org.apache.james.mailbox.store.mail.model.impl.SimpleProperty;
 import org.apache.james.util.CompletableFutureUtil;
 import org.apache.james.util.FluentFutureStream;
@@ -87,9 +80,6 @@ import com.datastax.driver.core.Session;
 import com.datastax.driver.core.UDTValue;
 import com.datastax.driver.core.querybuilder.QueryBuilder;
 import com.google.common.annotations.VisibleForTesting;
-import com.google.common.base.MoreObjects;
-import com.google.common.base.Objects;
-import com.google.common.base.Preconditions;
 import com.google.common.io.ByteStreams;
 import com.google.common.primitives.Bytes;
 
@@ -224,7 +214,7 @@ public class CassandraMessageDAO {
             .setUUID(MESSAGE_ID, cassandraMessageId.get()));
     }
 
-    private Pair<MessageWithoutAttachment, Stream<MessageAttachmentRepresentation>> message(Row row,ComposedMessageIdWithMetaData messageIdWithMetaData, FetchType fetchType) {
+    private Pair<MessageWithoutAttachment, Stream<MessageAttachmentRepresentation>> message(Row row, ComposedMessageIdWithMetaData messageIdWithMetaData, FetchType fetchType) {
         ComposedMessageId messageId = messageIdWithMetaData.getComposedMessageId();
 
         MessageWithoutAttachment messageWithoutAttachment =
@@ -324,161 +314,5 @@ public class CassandraMessageDAO {
         byte[] headerContent = new byte[row.getBytes(field).remaining()];
         row.getBytes(field).get(headerContent);
         return headerContent;
-    }
-
-    public static class MessageAttachmentRepresentation {
-
-        public static Builder builder() {
-            return new Builder();
-        }
-
-        public static class Builder {
-
-            private AttachmentId attachmentId;
-            private Optional<String> name;
-            private Optional<Cid> cid;
-            private Optional<Boolean> isInline;
-
-            private Builder() {
-                name = Optional.empty();
-                cid = Optional.empty();
-                isInline = Optional.empty();
-            }
-
-            public Builder attachmentId(AttachmentId attachmentId) {
-                Preconditions.checkArgument(attachmentId != null);
-                this.attachmentId = attachmentId;
-                return this;
-            }
-
-            public Builder name(String name) {
-                this.name = Optional.ofNullable(name);
-                return this;
-            }
-
-            public Builder cid(Optional<Cid> cid) {
-                Preconditions.checkNotNull(cid);
-                this.cid = cid;
-                return this;
-            }
-
-
-            public Builder cid(Cid cid) {
-                this.cid = Optional.ofNullable(cid);
-                return this;
-            }
-
-            public Builder isInline(boolean isInline) {
-                this.isInline = Optional.of(isInline);
-                return this;
-            }
-
-            public MessageAttachmentRepresentation build() {
-                Preconditions.checkState(attachmentId != null, "'attachmentId' is mandatory");
-                boolean builtIsInLine = isInline.orElse(false);
-                return new MessageAttachmentRepresentation(attachmentId, name, cid, builtIsInLine);
-            }
-        }
-
-        private final AttachmentId attachmentId;
-        private final Optional<String> name;
-        private final Optional<Cid> cid;
-        private final boolean isInline;
-
-        @VisibleForTesting
-        MessageAttachmentRepresentation(AttachmentId attachmentId, Optional<String> name, Optional<Cid> cid, boolean isInline) {
-            this.attachmentId = attachmentId;
-            this.name = name;
-            this.cid = cid;
-            this.isInline = isInline;
-        }
-
-        public AttachmentId getAttachmentId() {
-            return attachmentId;
-        }
-
-        public Optional<String> getName() {
-            return name;
-        }
-
-        public Optional<Cid> getCid() {
-            return cid;
-        }
-
-        public boolean isInline() {
-            return isInline;
-        }
-
-        @Override
-        public boolean equals(Object obj) {
-            if (obj instanceof MessageAttachmentRepresentation) {
-                MessageAttachmentRepresentation other = (MessageAttachmentRepresentation) obj;
-                return Objects.equal(attachmentId, other.attachmentId)
-                    && Objects.equal(name, other.name)
-                    && Objects.equal(cid, other.cid)
-                    && Objects.equal(isInline, other.isInline);
-            }
-            return false;
-        }
-
-        @Override
-        public int hashCode() {
-            return Objects.hashCode(attachmentId, name, cid, isInline);
-        }
-
-        @Override
-        public String toString() {
-            return MoreObjects
-                    .toStringHelper(this)
-                    .add("attachmentId", attachmentId)
-                    .add("name", name)
-                    .add("cid", cid)
-                    .add("isInline", isInline)
-                    .toString();
-        }
-    }
-
-    static class MessageWithoutAttachment {
-
-        private final MessageId messageId;
-        private final Date internalDate;
-        private final Long size;
-        private final Integer boduSize;
-        private final SharedByteArrayInputStream content;
-        private final Flags flags;
-        private final PropertyBuilder propertyBuilder;
-        private final MailboxId mailboxId;
-        private final MessageUid messageUid;
-        private final long modSeq;
-
-        public MessageWithoutAttachment(MessageId messageId, Date internalDate, Long size, Integer boduSize, SharedByteArrayInputStream content,
-                                        Flags flags, PropertyBuilder propertyBuilder, MailboxId mailboxId, MessageUid messageUid, long modSeq) {
-            this.messageId = messageId;
-            this.internalDate = internalDate;
-            this.size = size;
-            this.boduSize = boduSize;
-            this.content = content;
-            this.flags = flags;
-            this.propertyBuilder = propertyBuilder;
-            this.mailboxId = mailboxId;
-            this.messageUid = messageUid;
-            this.modSeq = modSeq;
-        }
-
-        public SimpleMailboxMessage toMailboxMessage(List<MessageAttachment> attachments) {
-            SimpleMailboxMessage simpleMailboxMessage = new SimpleMailboxMessage(messageId, internalDate, size, boduSize,
-                content, flags, propertyBuilder, mailboxId, attachments);
-            simpleMailboxMessage.setUid(messageUid);
-            simpleMailboxMessage.setModSeq(modSeq);
-            return simpleMailboxMessage;
-        }
-
-        public MailboxId getMailboxId() {
-            return mailboxId;
-        }
-
-        public MessageId getMessageId() {
-            return messageId;
-        }
     }
 }
